requirements.txt
asgiref==3.7.2
attrs==23.1.0
autobahn==23.6.2
Automat==22.10.0
black==23.12.1
cffi==1.16.0
channels==4.0.0
click==8.1.7
colorama==0.4.6
constantly==23.10.4
cryptography==41.0.7
daphne==4.0.0
Django==5.0
django-cors-headers==4.3.1
django-debug-toolbar==4.2.0
django-filter==23.5
django-resized==1.0.2
django-seed==0.3.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
Faker==22.0.0
flake8==6.1.0
hyperlink==21.0.0
idna==3.6
incremental==22.10.0
mccabe==0.7.0
mypy-extensions==1.0.0
packaging==23.2
pathspec==0.12.1
Pillow==10.1.0
platformdirs==4.1.0
psycopg2==2.9.9
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycodestyle==2.11.1
pycparser==2.21
pyflakes==3.1.0
PyJWT==2.8.0
pyOpenSSL==23.3.0
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
service-identity==23.1.0
six==1.16.0
sqlparse==0.4.4
tomli==2.0.1
toposort==1.10
Twisted==23.10.0
twisted-iocpsupport==1.0.4
txaio==23.1.1
typing_extensions==4.9.0
tzdata==2023.3
zope.interface==6.1
